pyproject.toml
[tool.poetry]
name = "codeleap"
version = "0.1.0"
description = ""
authors = ["Gabriel Caldas <gcaldasl2000@gmail.com>"]
readme = "README.md"
packages = [{include = "codeleap", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
